Replace custom animation curve with Unity AnimationCurves.
Also use individual curves for each animated field. This is
preferable to compress data.

+++ Assets/MixedRealityToolkit.Services/InputSimulation/InputPlaybackBehaviour.cs
@@ -37,7 +37,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
                 return;
             }
 
-            double currentTime = playable.GetTime();
+            float currentTime = (float)playable.GetTime();
             InputAnimationUtils.ApplyInputAnimation(InputAnimation, currentTime);
         }
     }
